Moved the tests as well into the respective packages
