+++ thirdeye/thirdeye-pinot/src/test/java/org/apache/pinot/thirdeye/common/dimension/DimensionMapTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.pinot.thirdeye.api;
+package org.apache.pinot.thirdeye.common.dimension;
 
 import org.apache.pinot.thirdeye.common.dimension.DimensionMap;
 import org.testng.Assert;
