+++ thirdeye/thirdeye-pinot/src/test/java/org/apache/pinot/thirdeye/common/metric/MetricTimeSeriesTest.java
@@ -14,15 +14,12 @@
  * limitations under the License.
  */
 
-package org.apache.pinot.thirdeye.api;
+package org.apache.pinot.thirdeye.common.metric;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import org.apache.pinot.thirdeye.common.metric.MetricSchema;
-import org.apache.pinot.thirdeye.common.metric.MetricTimeSeries;
-import org.apache.pinot.thirdeye.common.metric.MetricType;
 import org.apache.pinot.thirdeye.common.time.TimeRange;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
